LPS-181743 Extract ProjectDetailsCard to its own component
